refactor: delete the C capture library, rgsp-host now uses rgsp-cedar
Removes the last of src/, include/, and rgsp-host/build.rs. Cargo now owns
the whole build graph, so the standing 'rebuild librgspcast.a first or you
silently ship the previous C' footgun is gone with it.

- rgsp-host/src/capture.rs is now a thin re-export of rgsp_cedar::capture.
- video.rs casts fps/bitrate to i32 at the Capture::open_scaled call site,
  matching the crate's signature (the C FFI shim used u32).
- Ports test_idr_cadence.c and test_reopen_leak.c to rgsp-cedar/tests, with
  a shared tests/common module for the device-serialising Mutex.
- Makefile: drops librgspcast.a, bin/rgsp-cast, and the three C test-binary
  recipes; pak and all no longer depend on the archive.

src/rgsp-cast-cli.c stays for now (still lists in include/exclude of the
Makefile's bin/rgsp-cast build, which is removed here since it can no
longer link); Task 10 removes the file itself.

+++ rgsp-host/src/video.rs
@@ -161,8 +161,8 @@ impl VideoStream {
             PANEL_HEIGHT,
             self.cfg.width,
             self.cfg.height,
-            self.cfg.fps,
-            bitrate,
+            self.cfg.fps as i32,
+            bitrate as i32,
         )
         .map_err(|e| anyhow!("Capture::open: {e}"))?;
